requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==4.2.11
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-rosetta==0.10.0
filelock==3.13.4
idna==3.7
pillow==10.3.0
platformdirs==4.2.0
polib==1.2.0
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
whitenoise==6.6.0
